encapsulated components
