+++ src/pages/Dashboard/components/TeamManagement/components/PermissionsModal/index.ts
@@ -1,0 +1,1 @@
+export {PermissionsModal} from './PermissionsModal';
