+++ src/pages/Dashboard/components/TeamManagement/components/TeamsList/TeamsList.tsx
@@ -1,0 +1,5 @@
+export const TeamsList = () => {
+    return (
+        <p>You have teams</p>
+    )
+}
