+++ src/pages/Dashboard/components/TeamManagement/components/TeamsList/index.ts
@@ -1,0 +1,1 @@
+export {TeamsList} from './TeamsList';
