+++ src/pages/Dashboard/components/TeamManagement/components/TeamsPermission/index.ts
@@ -1,0 +1,1 @@
+export {TeamsPermission} from './TeamsPermission';
