+++ src/pages/Dashboard/components/TeamManagement/components/index.ts
@@ -1,0 +1,3 @@
+export {TeamsList} from './TeamsList';
+export {TeamsPermission} from './TeamsPermission';
+export {PermissionsModal} from './PermissionsModal';
